server/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
PyJWT
coverage
mysql-connector-python
pymysql